ensure full backwards and forward compatibility
The client assumes that NewPlayerSnapshotToClientMessages will NOT
contain reliable timestamps unless explicitely told so by the server.

The server now unconditionally sends the MPTweak
"uncompress.reliable_timestamps,True" to tell clients that the timestamps
it sends are reliable.

Full compatibitility is achieved for all possible combinations of
"old" PlayerSnapshots, NewPlayersnapshots with unreliable timestamps,
NewplayerSnapshots with reliable timestamps on client and server,
but the improvements of this version will only be available when using
it on both client and server.