Linux: add support for glibc >=2.34
olmod.so was relying on the internal glibc function _dl_sym() being
publically exported by the glibc. However, this is an internal function
of glibc which never was part of the official API/ABI. In glibc-2.34,
the visibility of this function was changed to "hidden", making olmod.so
fail to load with
`symbol lookup error: ./olmod.so: undefined symbol: _dl_sym`.

As a replacement, this patch uses dlvsym(), which is part of the official
glibc API/ABI so it won't go away in the future. This is also available
since a very long time, so it will also work on older glibc versions.