Add multiplayer chat commands and player bans.
Commands can be sent to the server via the /COMMAND [argument] syntax. The
following commands are implemented:

 * /KICK <player>: Kick a player from the server. Player can reconnect.
 * /BAN <player>: Ban player from the server, prevents joining.
 * /KICKBAN <player>: Kick and Ban a player.
 * /ANNOY <player>: Annoy-Ban player: Player is put into spectator state
                    on server, and chat messages are suppressed.
 * /UNBAN [<player>]: Unban specific player or all banned players.
 * /UNANNOY [<player>]: Un-annoy a player or all annoy-banned players.
 * /START: Start the match immediately (when in lobby)
 * /END: End the match immediately
 * /GIVEPERM <player>: grant a player the chat command permissions.
 * /REVOKEPERM [<player>]: revoke chat command permissions from a player
                           or all players.
 * /STATUS: short info status about chat command and bans.
 * /TEST: <player>: Test player name selction.

Player names are matches as follows:
 * If the name matches completely, that player is selected
 * If the player name begins with the specified pattern, it is selected.
 * If the the pattern is contained in several player names, the one
   where it matches earlier in the string is selected
 * If the match position is the same for multiple players, the one wich
   has fewer non-matched characters in the rest of the string is
   selected
 * The <player> pattern can use ? as wildcard character, matching every
   possible character (including non-typable unicode chars).
You can use the /TEST command to check which player would be selected.

The person who initially created the match has always permission to use
the chat commands. Further players can be added by the creator using the
/GIVEPERM command.

The Bans and Permissions stay effective until a new game is created on
the server by a different person. If the same person creates another
match, or the match is created by another player with chat command
permissions, all bans and permissions are kept as before. In all other
cases, all bans and chat permissions are reset. However, banned players
are also prevented from new creating matches on the server. If they attempt
to do create a new game, bans and permissions are also not reset.

Players can always use the /STATUS command to check whether bans or chat
permissions are active. This command does not require command permissions.