Fix detection of match creator.
Overload doesn't store the ID of the player who created the match, but only
the player's name. The player ban system relied on this name to check if
a player is the match creator.

However, player names aren't necessarily unique, and banned players
impersonating the actual match creator could lead to undesired effects
(thanks to Elvis for bringing this to my attention).

This patch fixes these issues by keeping the id of the player
with the private match data when the lobby was created, and using that
to find the true match creator when the players enter the lobby.