Disable the rendering of the cockpits of opponent ships in multiplayer. They are completely occluded by the ship hull mesh, so they are never visible.
The performance gain from this is negligible, though.  I measured an
increase of < 1% with 4 opponent ships visible. However, my system seems
to be mostly CPU-bound, there _might_ be a slightly higher gain on other
systems.

The patch is fairly simple and it doesn't hurt, so I'm offering it for
inclusion into olmod. YMMV.